TP2/requirements.txt
aiohttp
beautifulsoup4
lxml
pillow
selenium
aiofiles
requests
playwright
pytest